changed precision to unsigned long in engine routines.  Also started 'clean, 'norm' implementations
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@6221 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/CCC.cpp
@@ -11,7 +11,7 @@
 #include "gbring.hpp"
 #include "../d/M2mem.h"
 
-bool CCC::initialize_CCC(int prec) 
+bool CCC::initialize_CCC(unsigned long prec) 
 {
   initialize_ring(0);
   declare_field();
@@ -27,7 +27,7 @@ bool CCC::initialize_CCC(int prec)
   return true;
 }
 
-CCC *CCC::create(int prec)
+CCC *CCC::create(unsigned long prec)
 {
   CCC *result = new CCC;
   result->initialize_CCC(prec);
